change to use kulabyte web API
The API provided by Haivision does not allow the user to control channels created within the web interface; only channels created within the API itself. This version of the module uses a reverse-engineered version of the Web API instead, essentially emulating the web interface for control of channels within the web interface.